requirements.txt
pytest
qdrant-client
fastapi
uvicorn
requests
apscheduler
langchain-core
langchain-community
langchain-ollama
langchain-openai
openai
python-dotenv
psycopg2-binary
tqdm
